+++ packages/datadog-plugin-kafkajs/src/consumer.js
@@ -1,6 +1,7 @@
 'use strict'
 
 const dc = require('dc-polyfill')
+const { extract } = require('./utils')
 const { getMessageSize } = require('../../dd-trace/src/datastreams/processor')
 const { DsmPathwayCodec } = require('../../dd-trace/src/datastreams/pathway')
 const ConsumerPlugin = require('../../dd-trace/src/plugins/consumer')
@@ -90,25 +91,25 @@ class KafkajsConsumerPlugin extends ConsumerPlugin {
     }
   }
 
-  startBatch ({ topic, partition, messages, groupId }) {
-    if (this.config.dsmEnabled) {
-      this.tracer.setCheckpoint(['direction:in', `group:${groupId}`, `topic:${topic}`, 'type:kafka'])
-    }
-    this.startSpan({
-      resource: topic,
-      type: 'worker',
-      meta: {
-        component: 'kafkajs',
-        'kafka.topic': topic,
-        'kafka.message.offset': messages[0].offset,
-        'kafka.message.offset.last': messages[messages.length - 1].offset
-      },
-      metrics: {
-        'kafka.partition': partition,
-        'kafka.batch_size': messages.length
-      }
-    })
-  }
+  // startBatch ({ topic, partition, messages, groupId }) {
+  //   if (this.config.dsmEnabled) {
+  //     this.tracer.setCheckpoint(['direction:in', `group:${groupId}`, `topic:${topic}`, 'type:kafka'])
+  //   }
+  //   this.startSpan({
+  //     resource: topic,
+  //     type: 'worker',
+  //     meta: {
+  //       component: 'kafkajs',
+  //       'kafka.topic': topic,
+  //       'kafka.message.offset': messages[0].offset,
+  //       'kafka.message.offset.last': messages[messages.length - 1].offset
+  //     },
+  //     metrics: {
+  //       'kafka.partition': partition,
+  //       'kafka.batch_size': messages.length
+  //     }
+  //   })
+  // }
 
   finish () {
     if (beforeFinishCh.hasSubscribers) {
@@ -117,20 +118,6 @@ class KafkajsConsumerPlugin extends ConsumerPlugin {
 
     super.finish()
   }
-}
-
-function extract (tracer, bufferMap) {
-  if (!bufferMap) return null
-
-  const textMap = {}
-
-  for (const key of Object.keys(bufferMap)) {
-    if (bufferMap[key] === null || bufferMap[key] === undefined) continue
-
-    textMap[key] = bufferMap[key].toString()
-  }
-
-  return tracer.extract('text_map', textMap)
 }
 
 module.exports = KafkajsConsumerPlugin
