+++ packages/datadog-plugin-kafkajs/src/utils.js
@@ -1,0 +1,17 @@
+function extract (tracer, bufferMap) {
+  if (!bufferMap) return null
+
+  const textMap = {}
+
+  for (const key of Object.keys(bufferMap)) {
+    if (bufferMap[key] === null || bufferMap[key] === undefined) continue
+
+    textMap[key] = bufferMap[key].toString()
+  }
+
+  return tracer.extract('text_map', textMap)
+}
+
+module.exports = {
+  extract
+}
